zwischenstand

+++ src/main/java/controller/OConverter.java
@@ -3,7 +3,6 @@ package controller;
 import model.Datensatz;
 import model.Messwert;
 
-import java.util.ArrayList;
 
 public class OConverter {
 
@@ -18,11 +17,11 @@ public class OConverter {
      */
     public static String convertDatensatzToOutput(Datensatz datensatz) {
         StringBuilder sb = new StringBuilder();
-        sb.append("# FWHM = ").append(datensatz.getFWHM()).append(", ").append(datensatz.getIndexOf(datensatz.getL())).append(", ").append(datensatz.getIndexOf(datensatz.getR())).append("\n");
+        sb.append("# FWHM = ").append((float) datensatz.getFWHM()).append(", ").append(datensatz.getIndexOf(datensatz.getL())).append(", ").append(datensatz.getIndexOf(datensatz.getR())).append("\n");
         sb.append("# pos\tint\t\tenv\n");
         for (Messwert mw :
                 datensatz.getMesswertList()) {
-            sb.append(mw.getX()).append("\t").append(mw.getY()).append("\t\t").append(mw.getY_einhuellende()).append("\n");
+            sb.append((float) mw.getX()).append("\t").append((float) mw.getY()).append("\t\t").append((float) mw.getY_einhuellende()).append("\n");
         }
         return sb.toString();
     }
